backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
requests==2.31.0
Pillow>=10.4.0
python-dotenv==1.0.0
# Additional packages for enhanced image processing
opencv-python==4.8.1.78
numpy==1.24.4